Fix UserLockedForFailedAttempt for locked users (#8)
* Fix faillock for locked users

This fixes a bug getting the xyz.openbmc_project.User.Attributes
property UserLockedForFailedAttempt when a password is locked.
And this only happens when the xyz.openbmc_project.User.AccountPolicy
MaxLoginAttemptBeforeLockout property has a nonzero value (default 5).

This bug was introduced in 87a6ab71de880d5e28b95a2077cb9d2758e037d9
titled "Adapt to OpenBMC's Linux-PAM module changes".

This bug causes the phosphor Web app > Menu > Security and access > User
Management page to show no users, not allow users to be created, etc.

Also, when the faillock command fails with an error message, this
adds that error message to the journal.

Tested:
1. Locked a user (via the "login" command), and ensured that faillock
reported 5 failed attempts, and the D-Bus interface returned the
true value for the UserLockedForFailedAttempt property.
> busctl introspect xyz.openbmc_project.User.Manager \
>   /xyz/openbmc_project/user/admin

2. Removed the faillock file and observed the command failure.
```
> Change the faillock directory to a plain file:
> # mv /var/run/faillock /var/run/faillock-real
> # touch /var/run/faillock
> # faillock
> faillock: Error reading tally directory: Not a directory
> # busctl introspect xyz.openbmc_project.User.Manager
>   /xyz/openbmc_project/user/admin
> Failed to get all properties on interface
>   xyz.openbmc_project.User.Attributes: The operation failed internally.
> # journalctl (selected entries):
> phosphor-user-manager: /usr/sbin/faillock: Error opening the tally file
>   for admin:Not a directory
> phosphor-user-manager: The operation failed internally.
> Then to recover:
> # rm /var/run/faillock
> # mv /var/run/faillock-real /var/run/faillock
```

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>

* Updated log entry from faillock errors

This updates the log message sent when the faillock command fails to use
the ERROR keyword.
This also enhances the comments which explain the output from the
faillock command when it fails with an error.