Add ssh users to hostconsoleaccess group (#17)
* Add ssh users to hostconsoleaccess group

Users who are in the Phosphor "ssh" group are intended to have access to
the host console via SSH port 2200.  Users in the "ssh" phosphor
privilege group are added to the "hostconsoleaccess" Linux group.

Tested:
1. Create new user in the "ssh" group and ensure they are in the
   hostconsoleaccess Linux group and can use ssh port 2200.
2. Change a user to the "ssh" group and ensure they are in the
   hostconsoleaccess group and can use ssh.
3. Remove user from the "ssh" group and ensure they are removed from
   the hostconsoleaccess group and cannot ssh.
4. Create new user not in the "ssh" group and ensure they are not in
   the hostconsoleaccess Linux group and cannot use ssh port 2200.

Use journalctl --follow to monitor.  Look for:
PAM password auth succeeded for 'admin'
PAM password auth succeeded for 'newuser2'
Logins are restricted to the group hostconsoleaccess but user 'newuser'
is not a member
grep newuser /etc/passwd verify the shell /bin/nologin or /bin/sh

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>

* removed trailing whitespace