Tie SSH access to privilege role (#18)
This changes the create and update operations so the "ssh" privilege group
input is ignored, and the user's "ssh" privilege is determined by their
privilege role.  Only admin and operator users will have ssh, and readonly
users will not have ssh.

Tested:
Tested only by Redfish APIs and not by ipmitool.
Creating new users gives the correct access to host console.
Changing user account role gives the correct access to host console (that
part of this fix works), but reports the previous HostConsole AccountType
(this is a bug).  For someone to see this bug, a BMC admin would have to
change a user account's role, then use the Redfish API to GET that account
and look at the AccountType value.

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>