Service User: Fix stale data reading for service user
There are cases where service user is enabled for MFA using old driver
and the dbus properties are persisted. In the atest driver the service
user is disabled for MFA but the persisted data from old driver will
make the service user enabled for MFA.

This commit will fix this issue.

Signed-off-by: Abhilash Raju <abhilash.kollam@gmail.com>